Updated summary reports to contain all tables, and updated graphs and csv to be compatible with different runs

+++ most_up_to_date/ev100_post_processing.py
@@ -38,8 +38,9 @@ class PostProcess():
             sn_list = []
             sn_files = {}
             path = glob.glob(dlog_dir + "\\*\\*\\")
-            freq_mode = re.search("(.*)(\\\)(.*)(\\\)$", path[0]).group(3)
-            output_dir = os.path.join(output_dir, freq_mode, run)
+
+            #freq_mode = re.search("(.*)(\\\)(.*)(\\\)$", path[0]).group(3)
+            output_dir = os.path.join(output_dir)
 
             self.create_folder(output_dir)
 
@@ -72,14 +73,21 @@ class PostProcess():
             regex = "(tk_atpg_)(.*)"
             df_data["DFT Type"] = df_data.apply(lambda x: re.search(regex, x["Pattern Name"]).group(2).split("_")[0].upper(),
                                                 axis=1)
-            df_data["freq mode"] = df_data.apply(lambda x: freq_mode, axis=1)
+            df_data["freq mode"] = df_data.apply(lambda x: "SVS" if re.search("svs_", x["Pattern Name"])
+            else "TUR" if re.search("tur_", x["Pattern Name"])
+            else "SVSD1" if re.search("svsd1_", x["Pattern Name"])
+            else "NOM", axis=1)
 
             df_data["# failing"] = sn_df.notnull().sum(axis=1)
             df_data["# failing"].fillna("0", inplace = True)
             df_data = pd.concat([df_data, sn_df], axis = 1)
             df_data.pop('Index')
-            output_file = os.path.join(output_dir, freq_mode + "_postprocess_failures.csv")
-            df_data.to_csv(output_file, index=False, sep=',', header=True, mode='w')
+            output_file = os.path.join(output_dir, "postprocess_failures.csv")
+            if os.path.exists(output_file):
+                df_data.to_csv(output_file, index=False, sep=',', header=False, mode='a')
+            else:
+                df_data.to_csv(output_file, index=False, sep=',', header=True, mode='w')
+
 
 
     def create_folder(self, dir):
@@ -152,14 +160,15 @@ class PostProcess():
         plt.savefig(output_plot)
 
     def all_data_compiled(self, output_dir):
-        paths = glob.glob(output_dir + '\*\*\*.csv')
+        paths = glob.glob(output_dir + '\*.csv')
         passing_rate = {"INT": [], "SAF": [], "TDF": []}
-        freq_modes = []
-        for output_path in paths:
-            df_data = pd.read_csv(output_path)
-            output_path = re.search("(.*)(\\\)(.*)$", output_path).group(1)
-            freq_mode = re.search("(.*)(\\\)(.*)(\\\)(.*)$", output_path).group(3)
-            run = re.search("(.*)(\\\)(.*)$", output_path).group(3)
+        failing_vectors = []
+        output_path = paths[0]
+        df_map = pd.read_csv(output_path)
+        freq_modes = df_map['freq mode'].unique()
+        dft_options = ["INT", "SAF", "TDF"]
+        for freq_mode in freq_modes:
+            df_data = df_map[df_map["freq mode"].str.match(freq_mode)]
             for dft_type in passing_rate.keys():
                 total_chips = 0
                 passing = 0
@@ -169,6 +178,7 @@ class PostProcess():
                 df_vector_type.pop('DFT Type')
                 df_vector_type.pop('freq mode')
                 df_vector_type.pop('# failing')
+                failing_vectors.append(df_vector_type.notnull().sum(axis = 0).tolist())
                 for col in df_vector_type:
                     total_chips += 1
                     if df_vector_type[col].isnull().sum() == df_vector_type.shape[0]:
@@ -197,17 +207,28 @@ class PostProcess():
                     rate_list.append(0.0)
                     passing_rate[dft_type] = rate_list
 
-                if freq_mode not in freq_modes:
-                    freq_modes.append(freq_mode)
         total_parts = copy.deepcopy(list(passing_rate.values()))
         for i in range(len(total_parts)):
             for j in range(len(total_parts[0])):
                 total_parts[i][j] = int((total_parts[i][j] / 100.00) * total_chips)
 
+        df_failing_vectors = pd.DataFrame(failing_vectors, columns=list(df_vector_type))
+        df_failing_vectors.insert(loc=0, column='freq mode', value = "")
+        df_failing_vectors.insert(loc=1, column='DFT Type', value = "")
+        len_rows = df_failing_vectors.shape[0]
+        len_type = len(dft_options)
+
+        for r in range(int(len_rows/len_type)):
+            for j in range(len_type):
+                df_failing_vectors.iloc[r*len_type + j,1] = dft_options[j]
+                df_failing_vectors.iloc[r * len_type + j, 0] = freq_modes[r]
+
+
         df_passing_rate = pd.DataFrame(np.array(list(passing_rate.values())).T.tolist(),
                                        columns= list(passing_rate.keys()), index = freq_modes)
         df_parts_passing = pd.DataFrame(np.array(total_parts).T.tolist(),
                                        columns=list(passing_rate.keys()), index=freq_modes)
+        df_parts_passing["total #"] = total_chips
 
         summary_data = os.path.join(output_dir, "_".join(freq_modes) + " summary data.csv")
         with open(summary_data, 'w', newline="") as f:
@@ -220,6 +241,12 @@ class PostProcess():
             writer.writerow(["# of Parts Passing"])
         df_parts_passing.to_csv(summary_data, index=True, sep=',', header=True, mode='a')
 
+        with open(summary_data, 'a', newline="") as f:
+            writer = csv.writer(f)
+            writer.writerow("")
+            writer.writerow(["# of Vectors Failed by Part"])
+        df_failing_vectors.to_csv(summary_data, index=False, sep=',', header=True, mode='a')
+
         title = "_".join(freq_modes) + " " + " Passing Percentage"
         plt.title(title)
 
@@ -231,7 +258,7 @@ class PostProcess():
         plt.xlabel('Pattern Category')
         plt.ylabel('Percentage Passing (%)')
         plt.legend()
-        spacing = [-.2, .8]
+        spacing = [-.2, .8, 1.8, 2.8]
         self.add_labels(list(passing_rate.values()), spacing, True)
 
         plt.ylim(0, 100)
@@ -253,7 +280,7 @@ class PostProcess():
         plt.xlabel('Pattern Category')
         plt.ylabel('# of Passing Parts')
         plt.legend()
-        spacing = [-.2, .8]
+        spacing = [-.2, .8, 1.8, 2.8]
         self.add_labels(total_parts, spacing)
 
         output_plot = os.path.join(output_dir, title + '.jpg')
@@ -263,9 +290,9 @@ class PostProcess():
         for value in data:
             for i in range(len(value)):
                 if percent:
-                    plt.text(pos[i], value[i], str(round(value[i], 2)) + "%", ha='center', fontweight='bold')
+                    plt.text(pos[i], value[i], str(round(value[i], 2)) + "%", ha='center', fontweight='bold', fontsize = "small")
                 else:
-                    plt.text(pos[i], value[i], str(round(value[i], 2)), ha='center', fontweight = 'bold')
+                    plt.text(pos[i], value[i], str(round(value[i], 2)), ha='center', fontweight = 'bold', fontsize = "small")
                 pos[i] = pos[i] + .2
 
 
@@ -278,11 +305,12 @@ def main():
     # run = "dft_run_2021-07-07"
     # output_dir = r"G:\ATPG_CDP\pattern_execution\output"
     base_dir = r"C:\Users\rpenmatc\OneDrive - Qualcomm\Desktop"
-    runs = ["dft_run_2021-07-07"]
+    runs = ["dft_run_2021-07-01"]
     output_dir = r"C:\Users\rpenmatc\OneDrive - Qualcomm\Desktop\pattern_execution\output"
     post = PostProcess()
-    #post.dlog_csv_post_process(base_dir, runs, output_dir)
+    post.dlog_csv_post_process(base_dir, runs, output_dir)
    # post.passing_rate_graph(output_file)
+    #freq_modes = ["SVS", "NOM", "TUR", "SVSD1"]
     post.all_data_compiled(output_dir)
 
 if __name__ == "__main__":
